criando endpoints
