+++ semana25/amaro-back-end-challenge/package.json
@@ -19,10 +19,12 @@
     },
     "dependencies": {
         "@types/knex": "^0.16.1",
+        "@types/uuid": "^8.3.4",
         "cors": "^2.8.5",
         "dotenv": "^16.0.0",
         "express": "^4.17.2",
         "knex": "^1.0.2",
-        "mysql": "^2.18.1"
+        "mysql": "^2.18.1",
+        "uuid": "^8.3.2"
     }
 }
