+++ semana25/amaro-back-end-challenge/src/endpoints/getProductsByid.ts
@@ -1,0 +1,1 @@
+import { Request, Response } from "express";
