+++ semana25/amaro-back-end-challenge/src/index.ts
@@ -1,17 +1,5 @@
-import express, { Express } from "express";
-import cors from "cors";
-import { AddressInfo } from "net";
+import { app } from "./controller/app";
 
-const app: Express = express();
-
-app.use(express.json());
-app.use(cors());
-
-const server = app.listen(process.env.PORT || 3003, () => {
-    if (server) {
-        const address = server.address() as AddressInfo;
-        console.log(`Server is running in http://localhost:${address.port}`);
-    } else {
-        console.error(`Failure upon starting server.`);
-    }
-});
+app.get("/products/id");
+app.get("/products/name");
+app.get("/products/tags");
